// types/DadosRecibo.ts
export interface ReciboDados {
    anoCalendario: string;
    codigoIdentificador: string;
    nomeMuseu: string;
    logradouro: string;
    numero: string;
    complemento?: string;
    bairro: string;
    cep: string;
    municipio: string;
    uf: string;
    nomeDeclarante: string;
    horaData: string;
    numeroRecibo: string;
    totalBensDeclarados: string;
    bensMuseologicos: string;
    bensBibliograficos: string;
    bensArquivisticos: string;
    tipoDeclaracao: string;
  }
  